% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmosTrans.R
\name{atmosTrans}
\alias{atmosTrans}
\title{Broadband atmospheric transmissivity}
\usage{
atmosTrans(Pa, Kt = 1, theta, W)
}
\arguments{
\item{Pa}{\code{numeric} or \code{Raster*} object, atmospheric pressure (kPa).}

\item{Kt}{\code{numeric}, turbidity coefficient, defaults to \code{1.0} for
clean air (Allen \emph{et al.}, 1998).}

\item{theta}{\code{numeric} or \code{Raster*} object, solar zenith angle over
planar surfaces (radians).}

\item{W}{\code{numeric} or \code{Raster*} object, water in the atmosphere (mm).}
}
\description{
Broadband atmospheric transmissivity outlined in Hwang \emph{et al.} (2013).
}
\references{
Allen RG, Pereira LS, Raes D, Smith M (1998) \emph{FAO Irrigation and
Drainage Paper} 56. Food and Agriculture Organization of the United Nations
(FAO), Rome. Available online at
\url{https://appgeodb.nancy.inra.fr/biljou/pdf/Allen_FAO1998.pdf} (accessed
2016-03-07).

Hwang K, Choi M, Lee SO, Seo JW (2013) Estimation of instantaneous and daily
net radiation from MODIS data under clear sky conditions: a case study in
East Asia. \emph{Irrigation Science} 31, 1173-1184,
doi:10.1007/s00271-012-0396-3.
}
