% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evapoIntercept.R
\name{evapoIntercept}
\alias{evapoIntercept}
\alias{surfaceWetness}
\title{Interception evaporation}
\usage{
evapoIntercept(rH, alpha = 1.26, delta, gamma = 0.066, rnc)

surfaceWetness(rH)
}
\arguments{
\item{rH}{Relative humidity in \%.}

\item{alpha}{\code{numeric}. Priestley and Taylor (1972) constant, see
\code{\link{potentialEvaporation}}.}

\item{delta}{Slope of the saturation vapor pressure curve in kPa / deg C,
typically derived from \code{\link{vaporPressureSat}}.}

\item{gamma}{\code{numeric}. Psychrometric constant, defaults to \code{0.066}
kPa / deg C as proposed by Fisher et al. (2008).}

\item{rnc}{Net radiation to the canopy in W^2 / m^2.}
}
\value{
Interception evaporation in W^2 / m^2.
}
\description{
Calculate the interception evaporation and related parameters.
}
\section{Functions}{
\itemize{
\item \code{surfaceWetness}: Relative surface wetness.
}}
\references{
Fisher JB, Tu KP, Baldocchi DD (2008) Global estimates of the
land--atmosphere water flux based on monthly AVHRR and ISLSCP-II data,
validated at 16 FLUXNET sites. Remote Sensing of Environment 112: 901--919,
doi:10.1016/j.rse.2007.06.025.

Priestley CHB, Taylor RJ (1972) On the assessment of surface heat flux and
evaporation using large scale parameters. Monthly Weather Review 100: 81--92,
doi:10.1175/1520-0493(1972)100<0081:OTAOSH>2.3.CO;2.
}
\seealso{
\code{\link{surfaceWetness}}, \code{\link{radNetCanopy}}.
}

