% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventScore.R
\name{eventScore}
\alias{eventScore}
\title{Score Calculation}
\usage{
eventScore(meth.tumor, meth.tumor.status, transcriptome,
  methylation.event = c("sil", "enh"))
}
\arguments{
\item{meth.tumor}{Processed tumor methylome data; beta values (the probe should have either sil or enh probes condition as specified in 'methNorSel' function. rows: probe index, columns: normal samples)}

\item{meth.tumor.status}{Processed tumor methylation status (this dataset should have the coresponding methylation status of the tumor samples. rows: probe index, columns: normal samples)}

\item{transcriptome}{Transcriptome dataset of the tumor samples (columns: tumor samples, Gene name (as specified in the probe index))}

\item{methylation.event}{methylation event type. Silencing = 'sil',Enhancing = 'enh'}
}
\value{
A list contains: 1. Score.report : the score reported on each event (probe methylation event), 2. transcriptome : the transcriptome data used in the analysis (all the data are match with methylome and methylome staus), 3. transcriptome.nor.dis : the ranked normally distributed transcriptome data, 4. meth.tumor : methylome data used in score calculation, 5. meth.tumor.status : methylation status data used in score calculation
}
\description{
This function calculates the silencing/enhancing methylation score; which evaluates the level of correlation between hyper/hypomethylation and gene expression
}
