% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDRcal.R
\name{FDRcal}
\alias{FDRcal}
\title{FDR Calculation}
\usage{
FDRcal(score.values, meth.tumor, meth.tumor.status, transcriptome,
  no.permutation = 100, methylation.event = c("sil", "enh"), seed = 100)
}
\arguments{
\item{score.values}{The score calculated for each probe (output of "eventScore" function)}

\item{meth.tumor}{methylome data used in score calculation}

\item{meth.tumor.status}{methylation status data used in score calculation}

\item{transcriptome}{the transcriptome data used for score calculation (ranked normally distributed transcriptome data)}

\item{no.permutation}{number of desired permutation(default=100)}

\item{methylation.event}{methylation event type. Silencing = 'sil',Enhancing = 'enh'}
}
\value{
A list contains: 1. FDR.res : table containing the FDR calculated for each observed score, 2. permutation.res : the results of (n=100) permutations + mean permutated score + observed scores, 3. score.cutoff : the suggested score cutoff: score > 1.5 & FDR < 0.1 & false score < 0.5
}
\description{
This function estimates the false discovery rate (FDR) of the calculated scores, we adopted the strategy previously proposed in Significance Analysis of Microarrays (SAM) by Tusher et al.It permutates (n=100) the transcriptome data.
}
