% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_pin_issues}
\alias{get_pin_issues}
\title{Get noted pin issues.}
\usage{
get_pin_issues(dbconn)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}
}
\value{
a vector of text strings as written in the database.
}
\description{
Get noted pin issues.
}
