% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_pinhts}
\alias{get_pinhts}
\title{pull in default pin_height_list.rds}
\usage{
get_pinhts()
}
\value{
rds object read into session
}
\description{
pull in default pin_height_list.rds
}
\examples{
pinhts <- get_pinhts()
}
