% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtmosProfParam.R
\name{getAtmosProfParam}
\alias{getAtmosProfParam}
\title{Extract atmospheric profiles}
\usage{
getAtmosProfParam(x, prm = "Surface_Pressure", ext = NULL, dsn = "",
  cores = 1L, verbose = FALSE)
}
\arguments{
\item{x}{\code{character}, one or multiple .hdf filename(s).}

\item{prm}{\code{character}, one or multiple SDS to be processed. See
\code{\link{getAtmosProfSds}} for valid layer names.}

\item{ext}{\code{Extent} or any other objects from which an \code{Extent} can
be derived (optional), passed on to \code{\link[raster]{crop}}.}

\item{dsn}{\code{character}, target folder for file output (optional). Note
that if not provided, all resulting images are stored 'in memory'.}

\item{cores}{\code{integer}, number of cores for parallel computing, see
\code{\link{getAtmosProfBbox}}.}

\item{verbose}{\code{logical}, determines whether to print additional
information to the console.}
}
\value{
A \code{list} of \code{length(x)}. Sub-entries per list entry
correspond to the target SDS layers specified in \code{prm}.
}
\description{
Extract and, to a limited extent, process scientific datasets (SDS) from the
MODIS Atmospheric Profiles product (MOD/MYD07_L2).
}
\seealso{
\code{\link{GDALinfo}}.
}

