% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtmosProfSds.R
\name{getAtmosProfSds}
\alias{getAtmosProfSds}
\title{Retrieve MOD/MYD07_L2 SDS names}
\usage{
getAtmosProfSds(x, type = c("raw", "name"))
}
\arguments{
\item{x}{\code{character}, one or multiple .hdf filename(s).}

\item{type}{\code{character}, type of information to be returned.
\code{"raw"} provides some additional information on the single layers (e.g.,
number of rows and columns, bands, and data type), whereas \code{"name"}
returns parameters only that serve as input for \code{\link{getAtmosProfParam}}.}
}
\value{
A \code{character} vector holding SDS information.
}
\description{
Retrieve names of the single scientific datasets (SDS) associated with the
MODIS Atmospheric Profiles product (MOD/MYD_L2).
}
\seealso{
\code{\link{GDALinfo}}
}

