% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSwathExtent.R
\name{getSwathExtent}
\alias{getSwathExtent}
\title{Retrieve MOD/MYD07_L2 bounding box}
\usage{
getSwathExtent(x, cores = 1L)
}
\arguments{
\item{x}{\code{character}, one or multiple .hdf filename(s).}

\item{cores}{\code{integer}, number of cores for parallel computing, useful
when processing large amounts of data.}
}
\value{
If \code{length(x) == 1}, a single \code{Extent} object, else a
\code{list} of \code{Extent} objects.
}
\description{
Retrieve bounding box from the MODIS Atmospheric Profiles product
(MOD/MYD07_L2).
}
\seealso{
\code{\link{GDALinfo}}.
}
