% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSwathSDS.R
\name{getSwathSDS}
\alias{getSwathSDS}
\title{Extract data layers from MODIS Swath products}
\usage{
getSwathSDS(x, prm, ext = NULL, template = NULL, method = "bilinear",
  dsn = "", cores = 1L, verbose = FALSE)
}
\arguments{
\item{x}{\code{character}, one or multiple .hdf filename(s).}

\item{prm}{\code{character}, one or multiple SDS to be processed. See
\code{\link{listSwathSDS}} for valid layer names.}

\item{ext}{\code{Extent} or any other objects from which an \code{Extent} can
be derived, passed on to \code{\link[raster]{crop}}. The required coordinate
reference system is EPSG:4326 (see
\url{http://spatialreference.org/ref/epsg/wgs-84/}). Note that although the
usage of this argument is optional, specifying a reference extent results in
considerable speed gains, particularly when a \code{template} is provided.}

\item{template}{\code{Raster*} template with parameters that the target SDS
should be resampled to, see \code{\link{resample}}.}

\item{method}{\code{character}, the resampling method passed on to
\code{\link{resample}} if \code{template} is specified.}

\item{dsn}{\code{character}, target folder for file output (optional). Note
that if not provided, all resulting images are stored 'in memory'.}

\item{cores}{\code{integer}, number of cores for parallel computing, see
\code{\link{getSwathExtent}}.}

\item{verbose}{\code{logical}, determines whether to print additional
information to the console.}
}
\value{
A \code{list} of \code{length(x)}. Sub-entries per list entry
correspond to the target SDS layers specified in \code{prm}.
}
\description{
Extract and process scientific datasets (SDS) from MODIS Swath products.
}
\seealso{
\code{\link{GDALinfo}}.
}
