% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwur.R
\name{lwur}
\alias{lwur}
\title{Longwave upward radiation}
\usage{
lwur(sigma = 5.67e-08, lst, bse, ...)
}
\arguments{
\item{sigma}{\code{numeric}, Stefan-Boltzmann constant.}

\item{lst}{\code{Raster} object, land surface temperature.}

\item{bse}{\code{Raster} object, broadband surface emissivity.}

\item{...}{Further arguments passed on to \code{\link{overlay}}.}
}
\description{
Longwave upward radiation model proposed by Allen \emph{et al.} (2007; see
References).
}
\references{
Allen RG, Tasumi M, Trezza R (2007) Satellite-Based Energy Balance for
Mapping Evapotranspiration with Internalized Calibration (METRIC) -- Model.
\emph{Journal of Irrigation and Drainage Engineering} 133(4), 395-406,
\url{http://dx.doi.org/10.1061/(ASCE)0733-9437(2007)133:4(395)}
(accessed 2016-03-07).
}

