% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methStatus.R
\name{methStatus}
\alias{methStatus}
\title{Methylation Status Determination}
\usage{
methStatus(normal.db, meth.tumor, methylation.event = c("sil", "enh"))
}
\arguments{
\item{normal.db}{selected normal-sample datasets (the probe should have either sil or enh probes condition as specified in 'methNorSel' function. rows: probe index, columns: normal samples)}

\item{meth.tumor}{tumor methylome datasets (columns: tumor samples, rows: probeIDs)}

\item{methylation.event}{methylation event type. Silencing = 'sil',Enhancing = 'enh'}
}
\value{
A list contains: 1. Normal samples methylome data, 2. Tumor samples methylome data, 3. Tumor samples methylation status, 4. Specific Beta distribution (based on each specific probes value), 5. Universal Beta estimation of normal-samples distribution (output of 'fitdistr' function)
}
\description{
This function evaluates the methylation status of tumor samples comparing to their normal condition.
}
