% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi.R
\name{ndvi}
\alias{ndvi}
\title{Normalized difference vegetation index}
\usage{
ndvi(red, nir, filename = "", ...)
}
\arguments{
\item{red}{\code{Raster*} object, red band.}

\item{nir}{\code{Raster*} object, near-infrared band.}

\item{filename}{Output filename (optional).}

\item{...}{Additional arguments passed on to \code{\link{writeRaster}}.}
}
\value{
The calculated NDVI as \code{Raster*} object.
}
\description{
Calculate the normalized difference vegetation index (NDVI; Tucker, 1979)
from red and near-infrared wavelengths.
}
\author{
Florian Detsch
}
\references{
Tucker, C.J. (1979). Red and photographic infrared linear combinations for
monitoring vegetation. \emph{Remote Sensing of Environment} 8, 127-150,
doi:10.1016/0034-4257(79)90013-0. Available online at
\url{http://www.sciencedirect.com/science/article/pii/0034425779900130}
(accessed on 2016-03-02).
}
\seealso{
\code{\link{savi}}.
}

