% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netRadiation.R
\name{netRadiation}
\alias{netRadiation}
\title{Net radiation}
\usage{
netRadiation(rsd, rsu, rld, es, rlu, ...)
}
\arguments{
\item{rsd}{\code{Raster*}, shortwave downward radiation.}

\item{rsu}{\code{Raster*}, shortwave upward radiation.}

\item{rld}{\code{Raster*}, longwave downward radiation.}

\item{es}{\code{Raster*}, broadband surface emissivity.}

\item{rlu}{\code{Raster*}, longwave upward radiation.}

\item{...}{Further arguments passed on to \code{\link{overlay}}.}
}
\value{
A \code{Raster*} object holding information about net radiation.
}
\description{
Compute the net radiation as the remainder of the 4-component land surface
energy budget (shortwave and longwave upward and downward radiation,
respectively).
}
\seealso{
\code{\link{swdr}}, \code{\link{lwdr}},
\code{\link{surfaceEmissivity}}, \code{\link{lwur}}.
}
