% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometricConstant.R
\name{psychrometricConstant}
\alias{psychrometricConstant}
\title{Psychrometric constant}
\usage{
psychrometricConstant(Pa)
}
\arguments{
\item{Pa}{\code{numeric} or \code{Raster*} object, atmospheric pressure
(kPa).}
}
\value{
Same as input (kPa).
}
\description{
Calculate the psychrometric constant from the atmospheric pressure.
}
\references{
Allen RG, Pereira LS, Raes D, Smith M (1998) Crop evaporation -
Guidelines for computing crop water requirements. \emph{FAO Irrigation and
Drainage Paper} 56. Food and Agriculture Organization of the United Nations
(FAO), Rome. Available online at
\url{https://appgeodb.nancy.inra.fr/biljou/pdf/Allen_FAO1998.pdf} (accessed
2016-03-07).
}

