% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radNetSoil.R
\name{radNetSoil}
\alias{radNetSoil}
\title{Net radiation to the soil}
\usage{
radNetSoil(rn, krn = 0.6, lai, filename = "", ...)
}
\arguments{
\item{rn}{Net radiation.}

\item{krn}{Extinction coefficient, defaults to \code{0.6} as proposed by
Impens and Lemeur (1969; see References).}

\item{lai}{Leaf area index.}

\item{filename}{Output filename (optional).}

\item{...}{Additional arguments passed on to \code{\link{writeRaster}}.}
}
\description{
Calculate net radiation to the soil from overall net radiation and leaf area
index.
}
\references{
Fisher JB, Tu KP, Baldocchi DD (2008) Global estimates of the
land--atmosphere water flux based on monthly AVHRR and ISLSCP-II data,
validated at 16 FLUXNET sites. Remote Sensing of Environment 112: 901--919,
doi:10.1016/j.rse.2007.06.025.

Impens I, Lemeur R (1969) Extinction of net radiation in different crop
canopies. Theoretical and Applied Climatology 17(4): 403--412,
doi:10.1007\%2FBF02243377.
}
\seealso{
\code{\link{netRadiation}}.
}
