% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset-package.R
\docType{package}
\name{reset-package}
\alias{reset-package}
\alias{resetpackage}
\title{Remote Sensing-Based Evapotranspiration}
\description{
Remote Sensing-Based Evapotranspiration
}
\details{
More about what it does (maybe more than one line)
}
\author{
Florian Detsch
}
\keyword{package}

