% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{RESET}
\usage{
reset(normal.db, meth.tumor, transcriptome, methylation.event = c("sil",
  "enh"), FDR.permutation.no = 100, seed = 100)
}
\arguments{
\item{normal.db}{normal methylation}

\item{meth.tumor}{tumor methylation}

\item{transcriptome}{gene expression data}

\item{methylation.event}{=c('sil','enh')}

\item{FDR.permutation.no}{=100}
}
\value{
A list contains all the info regarding the analysis

#@examples
#pipelineCore=function(normal.db,meth.tumor,transcriptome,methylation.event=c('sil','enh'),FDR.permutation.no=100)
}
\description{
This function runs the comprehensive RESET analysis on the provided datasets
}
