% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd.R
\name{rsd}
\alias{rsd}
\title{Shortwave downward radiation}
\usage{
rsd(g_sc = 1367, theta, d_r, tau_sw)
}
\arguments{
\item{g_sc}{\code{numeric}, solar constant.}

\item{theta}{Solar incidence angle.}

\item{d_r}{Inverse squared relative earth-sun distance.}

\item{tau_sw}{Broadband atmospheric transmissivity.}
}
\description{
Shortwave downward radiation model proposed by Allen \emph{et al.} (2007; see
References)
}
\references{
Allen, R.G., Tasumi, M. and R. Trezza (2007). Satellite-Based Energy Balance
for Mapping Evapotranspiration with Internalized Calibration (METRIC) --
Model. \emph{Journal of Irrigation and Drainage Engineering} 133(4), 395-406,
\url{http://dx.doi.org/10.1061/(ASCE)0733-9437(2007)133:4(395)}
(accessed 2016-03-07).
}

