% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savi.R
\name{savi}
\alias{savi}
\title{Soil-adjusted vegetation index}
\usage{
savi(red, nir, L = 0.5, cores = 1L, filename = "", ...)
}
\arguments{
\item{red}{\code{Raster*} object, red band.}

\item{nir}{\code{Raster*} object, near-infrared band.}

\item{L}{\code{numeric} canopy background adjustment factor.}

\item{cores}{Number of cores for parallel computing. If \code{cores > 1L},
a separate \code{filename} is required for each layer in \code{red} (and
\code{nir}).}

\item{filename}{Output filename (optional).}

\item{...}{Additional arguments passed on to \code{\link{writeRaster}}.}
}
\value{
The calculated SAVI as \code{Raster*} object.
}
\description{
Calculate the soil-adjusted vegetation index (SAVI; Huete, 1988) from red and
near-infrared wavelengths.
}
\author{
Florian Detsch
}
\references{
Huete, A.R. (1988). A Soil-Adjusted Vegetation Index (SAVI).
\emph{Remote Sensing of Environment} 25, 295-309,
doi:10.1016/0034-4257(88)90106-X. Available online at
\url{http://www.sciencedirect.com/science/article/pii/003442578890106X}
(accessed on 2016-03-02).
}
\seealso{
\code{\link{ndvi}}.
}

