% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_calc.R
\name{set_calc_changes}
\alias{set_calc_changes}
\title{set_calc.R}
\usage{
set_calc_changes(dataSET)
}
\arguments{
\item{dataSET}{}
}
\description{
set_calc family of functions intended use is to caluclate values using QAed
set and sa datasets from the set_get_ and set_check_ functions.
These functions take a tibble/dataset as input and return either a tibble/dataset, nested
tibble, or model objects that are used within the analysis.
}
\details{
Calculate incremental and cumulative changes in pin height. Both within and across SET readers
}
