% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_changes}
\alias{set_check_changes}
\title{set_check_changes}
\usage{
set_check_changes(
  dataSET,
  duration = "1 year",
  mm_change = 20,
  drop_rows = FALSE,
  issues = c("Hole", "hole", "mussel", "Holr", "Shell", "Mussel", "edge of hole",
    "hole next to mussel"),
  ...
)
}
\arguments{
\item{dataSET}{SET data as returned by set_get_sets.}

\item{duration}{character string of the time interval to use for threshold calculation; "1 year" default.
Leverages \code{lubridate} capability to parse duration so values such as "1 week" and "10 months" are acceptable.}

\item{mm_change}{Change in pin height over the duration provided. Values greater than this (within the specified
duration) will be given a flag. Note: the reported change in the data is converted to an absolute change to
capture both an increase and a decrease in the measured values.}

\item{drop_rows}{TRUE, drops rows that don't meet the flag criteria. To return the full dataset, with
an appended column of flag set to FALSE. Defaults to FALSE to protect unwanted removal.}
}
\value{
tibble of SET data that's been trimmed down to show only measures that were made that fell above the
the threshold passed in the function call.
}
\description{
Calculates the rate of change from provided change in time and change in pin height.
Used for flagging and catching potential errors. Replaces the more simple incremental change
approach with one that allows for larger gaps in the data. Perhaps also catches more potential
issues for readings that occurred more closely together.
}
