% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_doublereads}
\alias{set_check_doublereads}
\title{Check set data for potential biases in SET reader through a}
\usage{
set_check_doublereads(dataSET)
}
\arguments{
\item{dataSET}{Double read data, typically output from set_get_doublereads}
}
\value{

}
\description{
Used in conjunction with set_get_doublereads
}
