% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_measures}
\alias{set_check_measures}
\title{set_check_measures}
\usage{
set_check_measures(
  dataSET,
  issues = c("Hole", "hole", "mussel", "Holr", "Shell", "Mussel", "edge of hole",
    "hole next to mussel"),
  ...
)
}
\arguments{
\item{dataSET}{SET data from set_get_sets}
}
\value{
tibble containing SET data in long format
}
\description{
returns a list of measures that have issues indicated in the notes or has
suspicious incremental changes. These are individual measurements and differ from
set_check_pins which returns pin_id for flagging an entire set of pin reads for a station direction.
}
\examples{
NULL


}
