% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_pins}
\alias{set_check_pins}
\title{set_check_pins}
\usage{
set_check_pins(
  dataSET,
  issues = c("Hole", "hole", "mussel", "Holr", "Shell", "Mussel", "edge of hole",
    "hole next to mussel"),
  ...
)
}
\arguments{
\item{issues}{Defaults to some common note flags that have been used, (hole, mussle, mussel hole)}

\item{...}{additional character strings of notes to flag.}
}
\value{
tibble containing SA data in long format. This tibble must have a "Notes" column to operate proprerly.
}
\description{
returns a tibble of measures that had recorded 'issue' i.e.
mussel hole, or grass tuft. Allows the filtering of these pins
in analysis downstream in set_clean_pins().
}
\examples{


}
