% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_recorded_vals}
\alias{set_check_recorded_vals}
\title{set_check_recorded_vals}
\usage{
set_check_recorded_vals(dbconn)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db.}
}
\value{
tibble with a flag column appended that is used in filtering out
potential errant values.
}
\description{
This function requires a connection to the DB compared to the
other functions that take the SET data due to the methods used for converting
character values to numerics.
}
\examples{

}
