% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_widgets.R
\name{set_check_widget}
\alias{set_check_widget}
\title{set_check_widget
Opens a Shiny widget that allows for browsing potential data entry issues
or exclude outliers.}
\usage{
set_check_widget(dataSET)
}
\arguments{
\item{dataSET}{SET data set as provided by set_get_sets()}
}
\value{
dataframe of selected points.
}
\description{
set_check_widget
Opens a Shiny widget that allows for browsing potential data entry issues
or exclude outliers.
}
\examples{

}
