% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get}
\alias{set_get}
\title{Get SET-MH data from Access DB}
\usage{
set_get(db)
}
\arguments{
\item{db}{file url for backend data base.}
}
\description{
This function connects and returns SET-MH data from
}
\examples{
set_get('data/DB.gdb')
}
\keyword{SET-MH,}
