% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_accretions}
\alias{set_get_accretions}
\title{set_get_accretions}
\usage{
set_get_accretions(dbconn)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}
}
\value{
tibble containing SA data in long format
}
\description{
return a tidy, long form, tibble of Surface Accretion (SA) data.
}
\examples{
# ADD_EXAMPLES_HERE

}
