% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_DB}
\alias{set_get_DB}
\title{SET_get_* functions that fetch data from provided database.}
\usage{
set_get_DB(dbPath)
}
\arguments{
\item{dbPath}{file url for backend data base.}
}
\value{
an odbc connection object to a database to be used in piped operations for extracting data from DB. Primarily used in other functions.
}
\description{
Get connected to from Access DB from which
}
\details{
This function connects and returns information relating to the SET-MH database.
Note: There can (will) be compatibility issues with the bit-versions of R and Access.
The AccessDB that this function was designed around was built on a 32-bit version.
The R bit-version must match to successfully connect so can be an issue.
There is a work around available which allows both bit-versions to be installed side-by-side
more info here: https://knowledge.autodesk.com/support/autocad/learn-explore/caas/sfdcarticles/sfdcarticles/How-to-install-64-bit-Microsoft-Database-Drivers-alongside-32-bit-Microsoft-Office.html
}
\keyword{SET,}
\keyword{SET-MH,}
\keyword{database}
