% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_doublereads}
\alias{set_get_doublereads}
\title{Check data for double reads.
Investigate if and when a double read of a SET occurs. The result can be used to
make adjustments to data and}
\usage{
set_get_doublereads(dataSET)
}
\arguments{
\item{dataSET}{}
}
\value{
tibble with data containing double reads.
}
\description{
Check data for double reads.
Investigate if and when a double read of a SET occurs. The result can be used to
make adjustments to data and
}
