% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_equipment_tbl}
\alias{set_get_equipment_tbl}
\title{read in SET equipment file}
\usage{
set_get_equipment_tbl(filepath)
}
\arguments{
\item{filepath}{}
}
\value{
tibble for use in left_joins to the raw data. The field within this output will
provide values for a calculation that will adjust the raw measured values into a standard
height (NAVD88) that will be used in the \code{\link{set_get_absolute_heights}}
}
\description{
SET arm vertical offset (height from receiver) and pin lengths vary depending on
the equipment used. To manage for this and adjust as needed depending on the readers
equipment. Complicating matters, some equipment changes within a site's time series.
To accommodate this a separate table is read-in which contains the set arm height and
the pin lengths for each set equipment (with a unique equipment_id field) that will be
matched up to the reading through a joining process.
}
