% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_pinlengths}
\alias{set_get_pinlengths}
\title{set_get_pinlengths}
\usage{
set_get_pinlengths(pin_numb, pin_table)
}
\arguments{
\item{pin_numb}{numeric pin number most often passed in from an assigned column name.}

\item{pin_table}{named vector pin number and pin length combinations, name being numeric pin number.}
}
\value{
numeric pin length in mm
}
\description{
set_get_pinlengths
}
