% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_readers}
\alias{set_get_readers}
\title{set_get_readers}
\usage{
set_get_readers(dbconn)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}
}
\value{
tibble containing SET readers for projects.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
