% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_receiver_elevation}
\alias{set_get_receiver_elevation}
\title{set_get_receiver_elevation
Retreive deep SET-rod receiver height as measured at the top of the receiver. NAVD88 is standard.}
\usage{
set_get_receiver_elevation(dbconn, plotID)
}
\arguments{
\item{plotID}{}
}
\value{
numeric elevation, used for correcting raw SET measures to NAVD88 elevaitons.
}
\description{
set_get_receiver_elevation
Retreive deep SET-rod receiver height as measured at the top of the receiver. NAVD88 is standard.
}
