% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_samplingevents}
\alias{set_get_samplingevents}
\title{set_get_samplingevents}
\usage{
set_get_samplingevents(dbconn)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}
}
\value{
tibble containing sampling event information for SET readings
}
\description{
Retrieve SET reading events or sampling information.
}
\examples{
# ADD_EXAMPLES_HERE
}
