% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_sets}
\alias{set_get_sets}
\title{set_get_sets}
\usage{
set_get_sets(dbconn)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}
}
\value{
tibble containing SET data in long format
}
\description{
return a tidy, long form, tibble of raw SET measurement data.
}
\examples{
# ADD_EXAMPLES_HERE

}
