% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_to_setr}
\alias{set_to_setr}
\title{set data to SETr data.}
\usage{
set_to_setr(dataSET)
}
\arguments{
\item{dataSET}{tibble of SET data from reSET::set_get_sets}
}
\value{
renamed tibble for use in SETr packages calc_change_cumu and calc_change_incr functions.
}
\description{
set data to SETr data.
}
\examples{
set_to_setr(SET_data) \%>\% SETr::calc_change_incr() # returns list of dataframes of incremental change rates.
}
