% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceEmissivity.R
\docType{methods}
\name{surfaceEmissivity}
\alias{surfaceEmissivity}
\alias{surfaceEmissivity,numeric,numeric-method}
\alias{surfaceEmissivity,Raster,Raster-method}
\title{Broadband surface emissivity}
\usage{
\S4method{surfaceEmissivity}{numeric,numeric}(ndvi, lai)

\S4method{surfaceEmissivity}{Raster,Raster}(ndvi, lai, ...)
}
\arguments{
\item{ndvi}{\code{numeric} or \code{Raster*} object, NDVI.}

\item{lai}{\code{numeric} or \code{Raster*} object, LAI.}

\item{...}{In case of \code{Raster*} input, further arguments passed on to
\code{\link{overlay}}.}
}
\value{
The broadband surface emissivity in the same format as the input.
}
\description{
Compute the broadband surface emissivity from the normalized difference
vegetation index (NDVI) and leaf area index (LAI) as proposed by Tasumi
(2003; see References).
}
\examples{
surfaceEmissivity(ndvi = 0.5, lai = 2)

}
\references{
Tasumi (2003) Progress in operational estimation of regional
evapotranspiration using satellite imagery. PhD dissertation, University of
Idaho, Moscow, Idaho, USA.
}
\seealso{
\code{\link{ndvi}}, \code{\link{savi}}, \code{\link{lwur}},
\code{\link{overlay}}.
}
