% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swathInfo.R
\name{swathInfo}
\alias{swathInfo}
\alias{getSwathDateTime}
\alias{getSwathDate}
\alias{getSwathTime}
\title{Retrieve Information from MODIS Swath Products}
\usage{
getSwathDateTime(x, format, ...)

getSwathDate(x, as_date = TRUE)

getSwathTime(x)
}
\arguments{
\item{x}{\code{character}, one or multiple .hdf filename(s).}

\item{format}{\code{character}, target format passed on to
\code{\link{strftime}}.}

\item{...}{Further arguments passed on to underlying functions (see Value).}

\item{as_date}{\code{logical}, determines whether or not to return a
\code{Date} object.}
}
\value{
\itemize{
\item \code{getSwathDateTime}: if \code{format} is missing, a \code{POSIXlt}
object created from \code{\link{strptime}}; else a \code{character} object
from \code{\link{strftime}}.
\item \code{getSwathDate}: if \code{as_date = TRUE} (default), a \code{Date}
object created from \code{\link{as.Date}}; else a \code{character} object
with a leading 'A'.
\item \code{getSwathTime}: a \code{character} object in the form of '\%H\%S'.
}
}
\description{
Set of helper functions to retrieve information from MODIS swath products.
}
\section{Functions}{
\itemize{
\item \code{getSwathDateTime}: Return MODIS Swath scan date and time

\item \code{getSwathDate}: Return MODIS Swath scan date

\item \code{getSwathTime}: Return MODIS Swath scan time
}}

\examples{
x <- "MOD05_L2.A2013001.0835.006.2015064204640.hdf"

getSwathDate(x)                  # scan date as 'Date'
getSwathDate(x, as_date = FALSE) # scan date as 'character'

getSwathTime(x) # scan time

getSwathDateTime(x)                            # scan date and time as 'POSIXlt'
getSwathDateTime(x, format = "\%Y-\%m-\%d \%H:\%S") # scan date and time as 'character'

}
