% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swdr.R
\name{swdr}
\alias{swdr}
\title{Shortwave downward radiation}
\usage{
swdr(Gsc = 1367, theta, d, tau)
}
\arguments{
\item{Gsc}{\code{numeric}, solar constant.}

\item{theta}{Solar incidence angle.}

\item{d}{Inverse squared relative earth-sun distance.}

\item{tau}{Broadband atmospheric transmissivity, typically derived from
\code{\link{atmosphericTransmissivity}}.}
}
\description{
Shortwave downward radiation model proposed by Allen \emph{et al.} (2007; see
References)
}
\references{
Allen, R.G., Tasumi, M. and R. Trezza (2007). Satellite-Based Energy Balance
for Mapping Evapotranspiration with Internalized Calibration (METRIC) --
Model. \emph{Journal of Irrigation and Drainage Engineering} 133(4), 395-406,
\url{http://dx.doi.org/10.1061/(ASCE)0733-9437(2007)133:4(395)}
(accessed 2016-03-07).
}

