% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vaporPressure.R
\name{vaporPressure}
\alias{vaporPressure}
\alias{vaporPressureSat}
\title{(Saturation) vapor pressure}
\usage{
vaporPressure(ta, rh)

vaporPressureSat(ta, slope = FALSE)
}
\arguments{
\item{ta}{Air temperature in deg C.}

\item{rh}{Relative humidity in \%.}

\item{slope}{\code{logical}. If \code{TRUE}, the slope of the saturation
vapor pressure curve is returned rather than the actual saturation vapor
pressure.}
}
\value{
The actual (or saturation) vapor pressure in kPa or, if \code{slope = TRUE},
the slope of the saturation vapor pressure curve in kPa / deg C.
}
\description{
Calculate actual (Peng et al. 2006) and saturation vapor pressure (Allen and
Robinson 2007) from air temperature and, in the case of the former, relative
air humidity.
}
\section{Functions}{
\itemize{
\item \code{vaporPressureSat}: Saturation vapor pressure in kPa.
}}
\references{
Allen RG, Robinson CW (2007) Evapotranspiration and Consumptive Irrigation
Water Requirements for Idaho. University of Idaho Research and Extension
Center: Kimberly, Idaho, USA.

Dingman LS (2008) Physical Hydrology. Waveland Press Inc.: Long Grove,
Illinois, USA.

Jensen ME, Burman RD, Allen RG (1990) Evapotranspiration and Irrigation Water
Requirements. ASCE Manuals and Reports on Engineering Practices No. 70.
American Society of Civil Engineers: New York, New York, USA.

Murray FW (1967) On the Computation of Saturation Vapor Pressure. Journal of
Applied Meteorology 6: 203--204, doi:10.1175/1520-0450(1967)006<0203:OTCOSV>2.0.CO;2.

Peng G, Li J, Chen Y, Norizan AP, Tay L (2006) High-resolution Surface
Relative Humidity Computation Using MODIS Image in Peninsular Malaysia.
Chinese Geographical Science 16(3): 260--264, doi:10.1007/s11769-006-0260-6.

Tetens O (1930) Ueber einige meteorologische Begriffe. Zeitschrift fuer
Geophysik 6: 297--309.
}

