% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi.R
\name{vi}
\alias{vi}
\alias{ndvi}
\alias{savi}
\alias{wdrvi}
\title{Various vegetation indices}
\usage{
ndvi(red, nir, cores = 1L, filename = "", ...)

savi(red, nir, L = 0.5, cores = 1L, filename = "", ...)

wdrvi(red, nir, ndvi = NULL, alpha = 0.2, cores = 1L, filename = "",
  ...)
}
\arguments{
\item{red}{\code{Raster*} object, red band.}

\item{nir}{\code{Raster*} object, near-infrared band.}

\item{cores}{Number of cores for parallel computing. If \code{cores > 1L},
a separate \code{filename} is required for each layer in \code{red} (and
\code{nir}).}

\item{filename}{Output filename (optional).}

\item{...}{Additional arguments passed on to \code{\link{writeRaster}}.}

\item{L}{\code{numeric} canopy background adjustment factor, defaults to
\code{0.5}.}

\item{ndvi}{\code{Raster*} object, e.g., derived from \code{\link{ndvi}}. If
specified, arguments \code{red} and \code{nir} are automatically ignored.}

\item{alpha}{\code{numeric} weighting coefficient, defaults to \code{0.2}.}
}
\value{
The calculated VI as \code{Raster*} object.
}
\description{
Calculate various vegetation indices (VI) from red and near-infrared
wavelengths. Currently available methods include the Normalized Difference
Vegetation Index (Tucker 1979), the Soil-adjusted Vegetation Index (Huete
1988), and the Wide Dynamic Range Vegetation Index (Gitelson 2004).
}
\section{Functions}{
\itemize{
\item \code{ndvi}: Normalized Difference Vegetation Index (Tucker 1979)

\item \code{savi}: Soil-adjusted Vegetation Index (Huete 1988)

\item \code{wdrvi}: Wide Dynamic Range Vegetation Index (Gitelson 2004)
}}

\references{
Gitelson, A.A. (2004). Wide dynamic range vegetation index for remote
quantification of crop biophysical characteristics. \emph{Journal of Plant
Physiology} 161, 165-173, doi:10.1078/0176-1617-01176. Available online at
\url{http://www.sciencedirect.com/science/article/pii/S0176161704705726}
(accessed on 2016-07-22).

Huete, A.R. (1988). A Soil-Adjusted Vegetation Index (SAVI).
\emph{Remote Sensing of Environment} 25, 295-309,
doi:10.1016/0034-4257(88)90106-X. Available online at
\url{http://www.sciencedirect.com/science/article/pii/003442578890106X}
(accessed on 2016-03-02).

Tucker, C.J. (1979). Red and photographic infrared linear combinations for
monitoring vegetation. \emph{Remote Sensing of Environment} 8, 127-150,
doi:10.1016/0034-4257(79)90013-0. Available online at
\url{http://www.sciencedirect.com/science/article/pii/0034425779900130}
(accessed on 2016-03-02).
}
\author{
Florian Detsch
}
