% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reSET-package.R
\docType{package}
\name{reSET-package}
\alias{reSET}
\alias{reSET-package}
\title{reSET: Gather and analyze SET-MH data to assess marsh elevation trends}
\description{
View, manage, analyze and visualize SET-MH data. Specifically for use along side the Marsh Elevation Monitoring Database (v.2.94) developed by the National Park Services Inventory and Monitoring .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/afstarke/reSET}
  \item Report bugs at \url{https://github.com/afstarke/reSET/issues}
}

}
\author{
\strong{Maintainer}: Adam Starke \email{astarke@tnc.org}

}
\keyword{internal}
