\name{RESET-package}
\alias{RESET-package}
\docType{package}
\title{
Reconstruction Set Test
}
\description{
Implementation of Reconstruction Set Test (RESET), a method for single sample variable set testing based on randomized reduced rank reconstruction error.
}
\details{
\tabular{ll}{
Package: \tab RESET\cr
Type: \tab Package\cr
Version: \tab 0.1.1 \cr
Date: \tab 2023\cr
License: \tab GPL-2 \cr
}
}
\author{
H. Robert Frost
}
\references{
    \itemize{
    \item Frost, H. R. (2023). Reconstruction Set Test (RESET): a computationally efficient method for single sample gene set testing based on randomized reduced rank reconstruction error. biorXiv e-prints. doi: https://doi.org/10.1101/2023.04.03.535366
    }
}
\note{
This work was supported by the National Institutes of Health grants R35GM146586, R21CA253408, P20GM130454 and P30CA023108.
}
\keyword{ package }

