% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_measures}
\alias{set_check_measures}
\title{set_check_measures}
\usage{
set_check_measures(
  dataSET,
  issues = c("Hole", "hole", "mussel", "Holr", "Shell", "Mussel", "edge of hole",
    "hole next to mussel"),
  ...
)
}
\arguments{
\item{dataSET}{SET data from set_get_sets}
}
\value{
tibble containing SET data with columns
}
\description{
returns a tibble of set data that includes columns indicating potential issues with the measurements
made as indicated in the notes and a flag for any pins that had issues at any point in time as provided
by the set_check_pins function.
}
\examples{
NULL


}
