% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_measures}
\alias{set_check_measures}
\title{set_check_measures}
\usage{
set_check_measures(dataSET)
}
\arguments{
\item{dataSET}{SET data from set_get_sets}
}
\value{
tibble containing SET data in long format
}
\description{
returns a list of pins that have issues indicated in the notes or has
suspicious incremental changes. Enhanced further with set_check_change function.
}
\examples{
# ADD_EXAMPLES_HERE

}
