% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_measures}
\alias{set_check_measures}
\title{set_check_measures}
\usage{
set_check_measures(dataSET)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}
}
\value{
tibble containing SA data in long format
}
\description{
returns a list of pins that have .
}
\examples{
# ADD_EXAMPLES_HERE

}
