% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_notes}
\alias{set_check_notes}
\title{set_check_notes
Used to check what notes have been made which can indicate pins that may have measurement bias}
\usage{
set_check_notes(dataSET)
}
\arguments{
\item{dataSET}{SET dataset from get_set_sets()}
}
\description{
set_check_notes
Used to check what notes have been made which can indicate pins that may have measurement bias
}
\examples{

}
