% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_check.R
\name{set_check_pins}
\alias{set_check_pins}
\title{set_check_pins}
\usage{
set_check_pins(
  dataSET,
  issues = c("Hole", "hole", "mussel", "Holr", "Shell", "Mussel", "edge of hole",
    "hole next to mussel"),
  ...
)
}
\arguments{
\item{issues}{Defaults to some common note flags that have been used, (hole, mussel, mussel hole)}

\item{...}{additional character strings of notes to flag.}
}
\value{
tibble containing SET data in long format. This tibble must have a "Notes" column to operate properly.
}
\description{
returns a vector of pin_ID's that had recorded a note consisting of an
'issue' i.e. mussel hole, or grass tuft. These are pins that, at any
point in the data set, had an issue noted. This vector can be used in
filtering pins from inclusion in analysis.
}
\examples{
NULL


}
