% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_DB}
\alias{set_get_DB}
\title{SET_get functions that fetch data from provided database.}
\usage{
set_get_DB(dbPath)
}
\arguments{
\item{dbPath}{file url for backend data base.}
}
\value{
an odbc connection object to a database to be used in piped operations for extracting data from DB. Primarily used in other functions.
}
\description{
Get connected to from Access DB from which
}
\details{
This function connects and returns information relating to the SET-MH database
}
\keyword{SET,}
\keyword{SET-MH,}
\keyword{database}
