% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_receiver_elevations}
\alias{set_get_receiver_elevations}
\title{set_get_receiver_elevations}
\usage{
set_get_receiver_elevations(dbconn)
}
\arguments{
\item{plotID}{}
}
\value{
sf object of surveyed plots with Lat/Lon, NAVD88 elevations, date
surveyed. Used for mapping or correcting raw SET measures to NAVD88
elevations.
}
\description{
Retrieve deep SET-rod receiver heights as measured at the top of
the receiver. NAVD88.
}
