% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_get.R
\name{set_get_stations}
\alias{set_get_stations}
\title{set_get_stations}
\usage{
set_get_stations(dbconn, epsg = 4269)
}
\arguments{
\item{dbconn}{Connection to Database returned from set_get_db}

\item{epsg}{desired coordinate system of the output using epsg code.}
}
\value{
sf object containing SET-MH stations. Includes survey elevation data for sites collected.
}
\description{
Retrieve SET-MH stations from database as an sf object. conversion of coordinates on the fly
}
\examples{
# studyStations <- set_get_DB(path) \%>\% set_get_stations()
}
