% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_pinhts}
\alias{set_pinhts}
\title{Create pin height table (named vector) using prompts.}
\usage{
set_pinhts(n_pins = 9)
}
\arguments{
\item{n_pins}{}
}
\value{
a named vector as well as a .rds object written to your current
  working directory
}
\description{
This utility function creates a named vector representing the
  pin length with a name as the pin number. This table/vector is used in the
}
\examples{
pinHts <- set_pinhts()
pinHts_18 <- set_pinhts(n_pins = 18)

}
\seealso{



}
